Added support for "dynamic manifests" (expiring auth/fragment URLs).
Fragment URLs or auth can expire with time. In case (1) a fragment verification fails, or (2) a fragment download fails with 403 Forbidden, we assume the fragment URL provided by the manifest or the user-provided "auth" string has expired and redownload the manifest or execute a user-provided script to get the fresh URLs/auth string(s).